+++ src/App.jsx
@@ -1,5 +1,5 @@
 import {Outlet} from "react-router-dom";
-import {AuthContextProvider} from "./components/context/AuthContext";
+import {AuthContextProvider} from "./context/AuthContext";
 
 import Navbar from "./components/Navbar";
