+++ src/context/AuthContext.jsx
@@ -1,5 +1,5 @@
 import {createContext, useContext, useEffect, useState} from "react";
-import {onUserStateChange} from "../../api/auth";
+import {onUserStateChange} from "../api/auth";
 
 const AuthContext = createContext();
